pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.37.0",
    "pypdf>=5.0.0",
    "faiss-cpu>=1.8.0.post1",
    "tiktoken>=0.7.0",
]
